+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.4 (09/15/2025)
+
+- Add callable main() method for svg
+- parametrize pytests and test cases
+- Additional test data
+- separate test output to `out` directory
+
 ## v0.1.3 (07/07/2024)
 
 - Updates from developing usage in wingwalker-inkex